requirements.txt
google-adk>=1.5.0
litellm>=1.74.0
drain3>=0.9.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0 